refactor: get rid of separate copy_assets.py
